02 - Future
In this class, you learned how to:

Create a Future object:
We learned how to create our first Future-type object and saw that it does not follow the rules of natural order of other objects. Usually, the Future Object needs a time to receive the information, so Dart himself puts it at the end of the task execution queue in order to wait for the information to arrive and not delay dependent objects.

Produce an Asynchronous Function:
We create a function that receives the async naming and, within it, we use a Future.delayed() method by adding the term await and thus fostering the wait action on Future-type objects.

Implement waiting time in BrunaBot:
In our project, we added the BotClock() class and implemented a clock() method that creates a wait function based on seconds. We added to our BrunaBot a few seconds of waiting in some waiting phrases. Thus, we created the simulation that our BrunaBot is seeking information from an external database! This was necessary to understand the context in which Future-type objects are inserted.