03 - Streams
In this class, you learned how to:

Create Streams:
We saw how to create a Stream, an object that receives information for a long period of time, where this information comes at intervals! Usually Streams are already given to us by entities outside our code, but because we are not connected with a Database/External Server/API, we decided to create our own Stream.

Create a Subscriber:
To listen to the information that comes from our Stream, we need to create an object of the StreamSubscriber type that has the ability to listen to the Stream through the listen() method, in addition to being able to manipulate the information, treat errors and much more.

Pause, restart and cancel a Stream:
We discovered the power to manipulate our subscription in the Stream by deciding when to pause it to prevent new information from coming, then we found out how to restart it to resume the flow of information and even cancel it, so as not to spend more resources!