04 - Finalizing the project
In this class, you learned how to:

Add new intelligences to the Bot:
We learned how to create a new class that contains all the information regarding the theme of a new kakoBot intelligence, in the specific case, we learned to add good manners to the project, responding to common phrases such as Hi, Good evening, Thank you among others.

Identify a real application:
We saw how our project can be implemented in a mobile application and that its functionality fits very well these days in chat applications.